Move browser facing code to separate MVC.
This commit:

- Moves :submit and :show to WebAuthBrowser MVC
- Moves WebAuth to API.WebAuth
- Updates Test
- Add minor changes to submit form CSS and view
