+++ lib/hexpm_web/templates/web_auth_browser/show.html.eex
@@ -1,10 +1,10 @@
 <h2>Device Activation</h2>
 
-<%= form_for @conn, Routes.web_auth_path(Endpoint, :submit), fn f -> %>
+<%= form_for @conn, Routes.web_auth_browser_path(Endpoint, :submit), fn f -> %>
     <div class="form-group">
       <%= label f, :user_code, "Enter the code displayed on your device" %>
       <%= ViewHelpers.text_input f, :user_code %>
-      <%= error_tag :user_code, @error %>
+      <%= error_tag @error %>
     </div>
 
     <button type="submit" class="btn btn-primary">Continue</button>
